+++ bin/templates/icons.tsx
@@ -1,3 +1,6 @@
+/**
+ * THIS FILE IS GENERATED
+ */
 ${icons.map(icon => "import " + icon.capitalized + " from '" + icon.root + icon.name + "'").join('\n')}
 
 export {