+++ src/itemInfo/index.tsx
@@ -1,0 +1,31 @@
+import React from 'react'
+
+import Item from '_utils/item/index'
+
+interface ItemInfoProps {
+  mainInfo: string,
+  className?: Classcat.Class,
+  icon?: React.ReactNode,
+  mainTitle?: string,
+  tag?: string,
+}
+
+const ItemInfo = ({
+  mainInfo,
+  className,
+  mainTitle,
+  icon,
+  tag,
+}:ItemInfoProps) => {
+  return (
+    <Item
+      className={className}
+      leftTitle={mainTitle}
+      leftBody={mainInfo}
+      leftAddon={icon}
+      tag={tag}
+    />
+  )
+}
+
+export default ItemInfo